refactor(enodebd): package code for enodebd devices (#13581)
* refactor(agw): package code for enodebd devices


* resolve conflicts
