+++ lte/gateway/python/magma/enodebd/devices/freedomfi_one/BUILD.bazel
@@ -1,0 +1,31 @@
+# Copyright 2022 The Magma Authors.
+
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+load("@rules_python//python:defs.bzl", "py_library")
+
+py_library(
+    name = "freedomfi_one",
+    srcs = [
+        "__init__.py",
+        "data_model.py",
+        "impl.py",
+        "params.py",
+        "states.py",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//lte/gateway/python/magma/enodebd:dp_client",
+        "//lte/gateway/python/magma/enodebd/data_models:transform_for",
+        "//lte/gateway/python/magma/enodebd/state_machines:acs_state_utils",
+        "//lte/gateway/python/magma/enodebd/state_machines:enb_acs",
+        "//lte/gateway/python/magma/enodebd/state_machines:enb_acs_impl",
+    ],
+)
