+++ lte/gateway/python/magma/enodebd/tests/BUILD.bazel
@@ -62,8 +62,8 @@ pytest_test(
         "//dp/protos:cbsd_python_proto",
         "//lte/gateway/python/magma/enodebd/data_models:data_model",
         "//lte/gateway/python/magma/enodebd/device_config:configuration_init",
-        "//lte/gateway/python/magma/enodebd/devices:baicells_qrtb",
         "//lte/gateway/python/magma/enodebd/devices:device_utils",
+        "//lte/gateway/python/magma/enodebd/devices/baicells_qrtb",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:enb_acs_builder",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:enodeb_handler",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:tr069_msg_builder",
@@ -227,7 +227,7 @@ pytest_test(
         "//lte/gateway/python/magma/enodebd/device_config:configuration_init",
         "//lte/gateway/python/magma/enodebd/device_config:configuration_util",
         "//lte/gateway/python/magma/enodebd/devices:device_utils",
-        "//lte/gateway/python/magma/enodebd/devices:freedomfi_one",
+        "//lte/gateway/python/magma/enodebd/devices/freedomfi_one",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:config_builder",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:enb_acs_builder",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:enodeb_handler",
@@ -249,7 +249,7 @@ pytest_test(
     deps = [
         "//lte/gateway/python/magma/enodebd/data_models:data_model_parameters",
         "//lte/gateway/python/magma/enodebd/device_config:configuration_init",
-        "//lte/gateway/python/magma/enodebd/devices:baicells_qrtb",
+        "//lte/gateway/python/magma/enodebd/devices/baicells_qrtb",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:enb_acs_builder",
         "//lte/gateway/python/magma/enodebd/tests/test_utils:enodeb_handler",
         requirement("parameterized"),
